+++ tests/__init__.py
@@ -1,0 +1,8 @@
+from selenium import webdriver
+from webdriver_manager.chrome import ChromeDriverManager
+
+def setup_get_driver_details():
+    options = webdriver.ChromeOptions()
+    options.headless = True
+    driver = webdriver.Chrome(options=options, executable_path=ChromeDriverManager().install())
+    return driver
