+++ tests/test_web_scrapper_amazon.py
@@ -1,23 +1,16 @@
-from selenium import webdriver
-from webdriver_manager.chrome import ChromeDriverManager
 from ..code.web.scraper.scrap.amazon import get_url_amazon, scrap_amazon
-
-
-def setup_get_driver_details():
-    options = webdriver.ChromeOptions()
-    options.headless = True
-    driver = webdriver.Chrome(options=options, executable_path=ChromeDriverManager().install())
-    return driver
-
+from . import setup_get_driver_details
 
 def test_get_url_amazon_1():
     item_name = "2021 Apple 10 2 inch iPad Wi Fi"
-    assert get_url_amazon(item_name) == "https://www.amazon.com/s?k=2021+Apple+10+2+inch+iPad+Wi+Fi&ref=nb_sb_ss_ts-doa-p_3_5"
-
+    search_term= item_name.replace(' ', '+')
+    assert get_url_amazon(item_name) == f"https://www.amazon.com/s?k={search_term}&ref=nb_sb_ss_ts-doa-p_3_5"
 
 def test_get_url_amazon_2():
-    assert get_url_amazon("Brita Longlast Replacement Filters Dispensers") == "https://www.amazon.com/s?k=Brita+Longlast+Replacement+Filters+Dispensers&ref=nb_sb_ss_ts-doa-p_3_5"
+    item_name = "Brita Longlast Replacement Filters Dispensers"
 
+    search_term= item_name.replace(' ', '+')
+    assert get_url_amazon(item_name) == f"https://www.amazon.com/s?k={search_term}&ref=nb_sb_ss_ts-doa-p_3_5"
 
 def test_scrap_amazon():
     item_name = "W. Trends Sunset Twin-Size Metal Bunk Bed - Black"
